Final Updates made

+++ src/components/OurProducts/HomeLoan.jsx
@@ -231,11 +231,11 @@ const HomeLoan = () => {
                     type="number"
                     className="w-full px-3 py-2 border rounded-md focus:outline-none focus:ring-1 focus:ring-[#18ADA5] transition-all text-sm"
                     value={interestRate}
-                    onChange={(e) => setInterestRate(parseInt(e.target.value))}
+                    onChange={(e) => setInterestRate(parseFloat(e.target.value))}
                     onBlur={() => {
                       const min = 0;
                       const max = activeTab === "personal" ? 60 : 60;
-                      const parsed = parseInt(interestRate) || min;
+                      const parsed = parseFloat(interestRate) || min;
                       const clamped = Math.min(Math.max(parsed, min), max);
                       setInterestRate(clamped);
                     }}
